+++ Program.cs
@@ -1,47 +1,42 @@
 ﻿using Microsoft.Data.Sqlite;
 using LabManager.Database;
 using LabManager.Repositories;
+using LabManager.Models;
 
-var databaseConfig = new DatabaseConfig();
-
-var databaseSetup = new DatabaseSetup(databaseConfig);
-
-var computerRepository = new ComputerRepository(databaseConfig);
-
-
- // Routing
-var modelName = args[0];
-var modelAction = args[1];
-
-if (modelName == "Computer")
+internal class Program
 {
-    if(modelAction == "List")
-    { 
-        Console.WriteLine("List Computer"); 
-        foreach(var computer in computerRepository.GetAll())
+    private static void Main(string[] args)
+    {
+        var databaseConfig = new DatabaseConfig();
+
+        var databaseSetup = new DatabaseSetup(databaseConfig);
+
+        var computerRepository = new ComputerRepository(databaseConfig);
+
+
+        // Routing
+        var modelName = args[0];
+        var modelAction = args[1];
+
+        if (modelName == "Computer")
         {
-            Console.WriteLine("{0}, {1}, {2}", computer.Id, computer.Ram, computer.Processor);
+            if (modelAction == "List")
+            {
+                Console.WriteLine("List Computer");
+                foreach (var computer in computerRepository.GetAll())
+                {
+                    Console.WriteLine("{0}, {1}, {2}", computer.Id, computer.Ram, computer.Processor);
+                }
+            }
+            if (modelAction == "New")
+            {
+                var id = Convert.ToInt32(args[2]);
+                var ram = args[3];
+                var processor = args[4];
+
+                var computer = new Computer(id, ram, processor);
+                computerRepository.Save(computer);
+            }
         }
     }
-    if(modelAction == "New")
-    {
-        var id = Convert.ToInt32(args[2]);
-        var ram = args[3];
-        var processor = args[4];
-    
-        var connection = new SqliteConnection(databaseConfig.ConnectionString);
-        connection.Open();
-
-
-        var command = connection.CreateCommand();
-        command.CommandText = " INSERT INTO Computers VALUES($id, $ram, $processor)";
-        command.Parameters.AddWithValue("$id", id);
-        command.Parameters.AddWithValue("$ram", ram);
-        command.Parameters.AddWithValue("$processor", processor);
-        
-        command.ExecuteNonQuery();
-        connection.Close();
-
-    }
 }
-
